tg_bot/requirements.txt
pyTelegramBotAPI
telebot
python-docx
sentence_transformers
numpy
scikit-learn
einops
requests